requirements.txt
django
djangorestframework
django-extensions
coverage
pylint
httpie
